added functions (comment add,comment update)

+++ app/Http/Controllers/ArticleController.php
@@ -3,86 +3,82 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-use App\Article;
 use App\Http\Requests\ArticleStoreRequest;
+
 use App\Repositories\ArticleRepository;
 use App\Repositories\TagRepository;
+use App\Repositories\UserRepository;
+
 use App\Services\ArticleService;
 
+use Illuminate\Support\Facades\Auth;
 class ArticleController extends Controller
 {
     protected $articleService;
+    protected $articleRepository;
+    protected $tagRepository;
+    protected $userRepository;
 
-    public function __construct(ArticleRepository $articleRepository,TagRepository $tagRepository)
+    //constructor
+    //args (article repo,tag repo)
+    public function __construct(ArticleRepository $articleRepository,TagRepository $tagRepository,UserRepository $userRepository)
     {
-        $this->articleService=new ArticleService($articleRepository,$tagRepository);
+        $this->articleService=new ArticleService($articleRepository);
 
+        $this->userRepository=$userRepository;
+        $this->articleRepository=$articleRepository;
+        $this->tagRepository=$tagRepository;
+
+        $this->middleware('auth');
     }
 
-    public function index()
+    //get all paginated articles
+    public function index(Request $request)
     {
-        //
-        $articles=$this->articleService->getAll();
+        $articles=$this->articleRepository->all();
         return view('articles.index',compact(['articles']));
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-        return view('articles.new',['tags'=>$this->articleService->getallTags()]);
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(ArticleStoreRequest $request)
-    {
-           $article=$this->articleService->store($request);
-           return redirect('/')->with('message', 'New Article Added!');
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+    //get one article
     public function show($id)
     {
         //
-        $article=$this->articleService->get($id);
+        $article=$this->articleRepository->get($id);
         return view('articles.detail',['article'=>$article]);
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+    //store an article
+    public function store(ArticleStoreRequest $request)
+    {
+        //
+        $article=$this->articleService->store($request);
+        return redirect('/')->with('message', 'New Article Added!');
+    }
+    //create new article (view)
+    public function create()
+    {
+        //
+        return view('articles.new',['tags'=>$this->tagRepository->all()]);
+
+    }
+
+    //edit an article (view)
     public function edit($id)
     {
         //
-        $tags=$this->articleService->getallTags();
-        $article=$this->articleService->get($id);
-        return view('articles.new',compact('tags','article'));
+        $user=$this->userRepository->getCurrentUser();
+        $article=$this->articleRepository->get($id);
+        if($user->can('update',$this->articleRepository->get($id))){
+            $tags=$this->tagRepository->all();
+            $article=$this->articleRepository->get($id);
+            return view('articles.new',compact('tags','article'));
+        }else{
+            return redirect()->back()->with('message','you can only update your own post');
+        }
+
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+    //update article data
     public function update(ArticleStoreRequest $request, $id)
     {
         //
@@ -91,18 +87,15 @@ class ArticleController extends Controller
 
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+    //delete an article
     public function destroy($id)
     {
         //
+       if(Auth::user()->can('delete',$this->articleRepository->get($id)))
        if($this->articleService->delete($id)){
-           return redirect()->back()->with('message','Article Deleted');
+           return redirect('/')->with('message','Article Deleted');
        }
+       return redirect()->back()->with('message','not allowed to delete this article');
 
     }
 }
